+++ app/src/main/java/org/sopt/soptseminar/data/service/GithubServiceCreator.kt
@@ -1,14 +1,14 @@
-package org.sopt.soptseminar.data.api
+package org.sopt.soptseminar.data.service
 
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 
-object SoptServiceCreator {
-    private const val BASE_URL = "http://13.124.62.236/"
+object GithubServiceCreator {
+    private const val BASE_URL = "https://api.github.com/"
 
     private val retrofit: Retrofit = Retrofit.Builder().baseUrl(BASE_URL).addConverterFactory(
         GsonConverterFactory.create()
     ).build()
 
-    val soptService: SoptService = retrofit.create(SoptService::class.java)
+    val githubService: GithubService = retrofit.create(GithubService::class.java)
 }
